+++ src/SFA.DAS.Payments.Monitoring.Alerts.Function/SendSlackAlert.cs
@@ -3,8 +3,7 @@ using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
+using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.Payments.Monitoring.Alerts.Function.Services;
 
@@ -19,7 +18,7 @@ namespace SFA.DAS.Payments.Monitoring.Alerts.Function
             _slackService = slackService;
         }
 
-        [FunctionName("HttpTrigger1")]
+        [Function("HttpTrigger1")]
         public async Task<IActionResult> SendToChannelOne(
             [HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)] HttpRequest req, ILogger log)
         {
@@ -37,7 +36,7 @@ namespace SFA.DAS.Payments.Monitoring.Alerts.Function
             return new OkObjectResult("");
         }
 
-        [FunctionName("HttpTrigger2")]
+        [Function("HttpTrigger2")]
         public async Task<IActionResult> SendToChannelTwo(
             [HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)] HttpRequest req, ILogger log)
         {
